+++ gerenciador_postgres/gui/privileges_view.py
